make webgl canvas more resistant to small random noise
