+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "thumbmarkjs",
-  "version": "0.11.4",
+  "version": "0.11.5",
   "description": "",
   "main": "dist/index.js",
   "module": "dist/index.js",
